Fix duration measurement.
On Linux, Stopwatch.GetTimestamp returns a different value than on Linux, making it so that there are 2 different kind of tick counts.
Stopwatch also doesn't expose anything other than creating a new instance of Stopatch and accessing the .Elapsed.Ticks property (make sure to not use the .ElapsedTicks otherwise you'll have created an instance for nothing!) which will return the converted tick count according to the OS's clock frequency.